+++ app/src/main/java/com/example/mpi/ui/pilar/PilarActivity.kt
@@ -9,7 +9,9 @@ import com.example.mpi.ui.pilar.cadastroPilar
 import com.example.mpi.ui.pilar.EditarPilarActivity
 import com.example.mpi.data.DatabaseHelper
 import android.util.Log
+import androidx.activity.enableEdgeToEdge
 import com.example.mpi.data.Pilar
+import com.example.mpi.repository.PilarRepository
 
 
 class PilarActivity : AppCompatActivity() {
@@ -17,14 +19,17 @@ class PilarActivity : AppCompatActivity() {
     private lateinit var binding: ActivityPilarBinding
     private lateinit var dbHelper: DatabaseHelper
     private lateinit var pilarAdapter: PilarAdapter
+    private lateinit var pilarRepository: PilarRepository
     private val listaPilares = mutableListOf<Pilar>()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
         binding = ActivityPilarBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
         dbHelper = DatabaseHelper(this)
+        pilarRepository = PilarRepository.getInstance(this)
 
         ////////////////////// Carregando informações do usuário////////////////////////////////
         val intentExtra = intent
@@ -37,7 +42,7 @@ class PilarActivity : AppCompatActivity() {
         ////////////////////////////////////////////////////////////////////////////////
 
         binding.recyclerViewPilares.layoutManager = LinearLayoutManager(this)
-        pilarAdapter = PilarAdapter(listaPilares, { pilar -> editarPilar(pilar) }, { pilar -> excluirPilar(pilar) })
+        pilarAdapter = PilarAdapter(listaPilares, { pilar -> editarPilar(pilar) }, { pilar -> excluirPilar(pilar) }, this)
         binding.recyclerViewPilares.adapter = pilarAdapter
 
         carregarPilares()
@@ -70,7 +75,6 @@ class PilarActivity : AppCompatActivity() {
             DatabaseHelper.COLUMN_PILAR_DESCRICAO,
             DatabaseHelper.COLUMN_PILAR_DATA_INICIO,
             DatabaseHelper.COLUMN_PILAR_DATA_TERMINO,
-            DatabaseHelper.COLUMN_PILAR_IS_APROVADO,
             DatabaseHelper.COLUMN_PILAR_PERCENTUAL,
             DatabaseHelper.COLUMN_PILAR_ID_CALENDARIO,
             DatabaseHelper.COLUMN_PILAR_ID_USUARIO
@@ -93,7 +97,6 @@ class PilarActivity : AppCompatActivity() {
                 val descricao = getString(getColumnIndexOrThrow(DatabaseHelper.COLUMN_PILAR_DESCRICAO))
                 val dataInicio = getString(getColumnIndexOrThrow(DatabaseHelper.COLUMN_PILAR_DATA_INICIO))
                 val dataTermino = getString(getColumnIndexOrThrow(DatabaseHelper.COLUMN_PILAR_DATA_TERMINO))
-                val aprovado = getInt(getColumnIndexOrThrow(DatabaseHelper.COLUMN_PILAR_IS_APROVADO)) > 0
                 val percentual = getDouble(getColumnIndexOrThrow(DatabaseHelper.COLUMN_PILAR_PERCENTUAL))
                 val idCalendario = getInt(getColumnIndexOrThrow(DatabaseHelper.COLUMN_PILAR_ID_CALENDARIO))
                 val idUsuario = getInt(getColumnIndexOrThrow(DatabaseHelper.COLUMN_PILAR_ID_USUARIO))
@@ -105,7 +108,6 @@ class PilarActivity : AppCompatActivity() {
                         descricao,
                         dataInicio,
                         dataTermino,
-                        aprovado,
                         percentual,
                         idCalendario,
                         idUsuario
@@ -119,45 +121,49 @@ class PilarActivity : AppCompatActivity() {
     }
 
     private fun editarPilar(pilar: Pilar) {
-        val intent = Intent(this, EditarPilarActivity::class.java)
-        intent.putExtra("pilar_id", pilar.id)
-        intent.putExtra("pilar_nome", pilar.nome)
-        intent.putExtra("pilar_descricao", pilar.descricao)
-        intent.putExtra("pilar_data_inicio", pilar.dataInicio)
-        intent.putExtra("pilar_data_termino", pilar.dataTermino)
-        intent.putExtra("pilar_aprovado", pilar.aprovado)
-        intent.putExtra("pilar_percentual", pilar.percentual)
-        intent.putExtra("pilar_id_calendario", pilar.idCalendario)
-        intent.putExtra("pilar_id_usuario", pilar.idUsuario)
-        startActivity(intent)
-        android.widget.Toast.makeText(this, "Editar: ${pilar.nome}", android.widget.Toast.LENGTH_SHORT).show()
+            val intent = Intent(this, EditarPilarActivity::class.java)
+            intent.putExtra("pilar_id", pilar.id)
+            intent.putExtra("pilar_nome", pilar.nome)
+            intent.putExtra("pilar_descricao", pilar.descricao)
+            intent.putExtra("pilar_data_inicio", pilar.dataInicio)
+            intent.putExtra("pilar_data_termino", pilar.dataTermino)
+            intent.putExtra("pilar_percentual", pilar.percentual)
+            intent.putExtra("pilar_id_calendario", pilar.idCalendario)
+            intent.putExtra("pilar_id_usuario", pilar.idUsuario)
+            startActivity(intent)
+            android.widget.Toast.makeText(this, "Editar: ${pilar.nome}", android.widget.Toast.LENGTH_SHORT).show()
     }
 
     private fun excluirPilar(pilar: Pilar) {
-        val db = dbHelper.writableDatabase
-        val whereClause = "${DatabaseHelper.COLUMN_PILAR_ID} = ?"
-        val whereArgs = arrayOf(pilar.id.toString())
-        val deletedRows = db.delete(DatabaseHelper.TABLE_PILAR, whereClause, whereArgs)
-        if (deletedRows > 0) {
-            listaPilares.remove(pilar)
-            pilarAdapter.notifyDataSetChanged()
+        if(pilarRepository.validarExclusaoPilar(pilar) == true){
+            val db = dbHelper.writableDatabase
+            val whereClause = "${DatabaseHelper.COLUMN_PILAR_ID} = ?"
+            val whereArgs = arrayOf(pilar.id.toString())
+            val deletedRows = db.delete(DatabaseHelper.TABLE_PILAR, whereClause, whereArgs)
+            if (deletedRows > 0) {
+                listaPilares.remove(pilar)
+                pilarAdapter.notifyDataSetChanged()
 
-            //verificando se o Pilar era o único existente, se sim o registro da tabela calendário sera apagado
-            if (listaPilares.isEmpty()) {
-                val idCalendarioExcluir = pilar.idCalendario
-                val whereClauseCalendario = "${DatabaseHelper.COLUMN_CALENDARIO_ID} = ?"
-                val whereArgsCalendario = arrayOf(idCalendarioExcluir.toString())
-                db.delete(
-                    DatabaseHelper.TABLE_CALENDARIO,
-                    whereClauseCalendario,
-                    whereArgsCalendario
-                )
+                //verificando se o Pilar era o único existente, se sim o registro da tabela calendário sera apagado
+                if (listaPilares.isEmpty()) {
+                    val idCalendarioExcluir = pilar.idCalendario
+                    val whereClauseCalendario = "${DatabaseHelper.COLUMN_CALENDARIO_ID} = ?"
+                    val whereArgsCalendario = arrayOf(idCalendarioExcluir.toString())
+                    db.delete(
+                        DatabaseHelper.TABLE_CALENDARIO,
+                        whereClauseCalendario,
+                        whereArgsCalendario
+                    )
+                }
+
+                android.widget.Toast.makeText(this, "Pilar '${pilar.nome}' excluído com sucesso!", android.widget.Toast.LENGTH_SHORT).show()
+            } else {
+                android.widget.Toast.makeText(this, "Erro ao excluir o pilar.", android.widget.Toast.LENGTH_SHORT).show()
             }
-
-            android.widget.Toast.makeText(this, "Pilar '${pilar.nome}' excluído com sucesso!", android.widget.Toast.LENGTH_SHORT).show()
-        } else {
-            android.widget.Toast.makeText(this, "Erro ao excluir o pilar.", android.widget.Toast.LENGTH_SHORT).show()
+            db.close()
+        }else{
+            android.widget.Toast.makeText(this, "Erro! Existem subpilares ou ações existentes vinculadas ao pilar.", android.widget.Toast.LENGTH_SHORT).show()
         }
-        db.close()
+
     }
 }
